astra2/requirements.txt
pydirectinput
pygetwindow
Pillow
opencv-contrib-python
numpy
PyAutoGUI